add function to unfold percentage values
Percentage values (provided by a map containing the referenced dimension
(:width or :height) and the actual percentage value) inside of attribute
maps can be transformed into absolute values by providing a base height
and width (optimally from the root svg element) and calling the
functions. Two helper functions (one for determining if a key-pair is
applicable for percentage conversion and one for actually performing the
transformation on one key-value pair) are included.